backend/requirements.txt
asgiref==3.4.1
Django==3.2.7
django-modeltranslation==0.17.3
django-simple-history==3.0.0
djangorestframework==3.12.4
psycopg2==2.9.1
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
sqlparse==0.4.2
uritemplate==3.0.1
